[OPENBUS-2475] [Java] Assistente não pode aceitar tempo de espera entre tentativas igual a ZERO - Muda o tipo do intervalo do assistente de Integer para Float
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/sdk/java/branches/02_00_01@151104 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ core/src/tecgraf/openbus/assistant/AssistantParams.java
@@ -38,7 +38,7 @@ public class AssistantParams {
    * espera pelo menos o tempo indicado por esse parâmetro antes de tentar uma
    * nova tentativa.
    */
-  public Integer interval;
+  public Float interval;
   /**
    * O ORB a ser utilizado pelo assistente para realizar suas tarefas. O
    * assistente também configura esse ORB de forma que todas as chamadas feitas
